pyproject.toml
[project]
name = "medium-md-fetcher"
version = "0.1.0"
description = "A command-line tool and Python library for fetching Medium articles."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "diskcache>=5.6.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.0.0",
]

[dependency-groups]
dev = [
    "coverage>=7.9.2",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "ruff>=0.12.4",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.isort]
lines-after-imports = 2
combine-as-imports = true
